mingw: allow cross building MK for win64
This commit introduces new recipes for cross building MK, as a
library, and its dependencies, for building OONI for win64.

Limitations:

1. for now we're addressing only win64, which is anyway the kind
of build that we most likely want to have;

2. for unknown reasons, the `measurement_kit.exe` binary built
using this procedure is not statically linked, while, at the
same time, the libraries produced by this build are okay in my
system to obtain statically linked OONI binaries;

3. the reason I'm using in the keg-only is not very accurate.

I'll stop here, because this is already a great simplification
over the existing situation, and because my original goal was
to make the OONI build easier for us for Windows systems.

I'll open bugs for the above limitations.

While there, I've also improve measurement-kit.rb.