cross.rb: package with cross building scripts
I am increasingly comfortable with this Homebrew based set of
build scripts, because I can create bottles and publish them at
bintray. More generally, all the problems related to having to
maintain a distribution are levied by the fact that Homebrew is
already a distribution. So, I'm going to try and use it also for
building iOS packages and possibly Android.

This is the starting point: adding a tool that I can use to
invoke ./configure and get the desired behaviour. It seems it
is already working quite well for libmaxminddb.